Reserve pool space on copy construction
When copying a HashTable, we know that the pool will need to be
able to allocate exactly as many nodes as the 'size' of the
table being copied.  By requesting the pool reserve this many
nodes, we get a minor optimization that turns out to be very
important when performing exception-guarantee tests of the
assignment operator, which will repeatedly attempt the assignment
with a test allocator slowly raising the limit of available
memory until the operation can just succeed.  This optimization
causes the allocotor to fail early every time, until it has
successfully acquired enough memory to perform the rest of the
operation, rather than looping quadratically copying one more
list of nodes that is just one longer, until we finally get the
successful final assignment.

As this optimization speeds up test case 9, we restore the
regular running of its functionality against a wider range
of concerns for test types, more than consuming the freed
up time.

Finally, we provide a litte more macro hygiene, undefined
a popular test-case specific macro that was leaking out of
each test case.  This actually turned up a copy-paste bug,
so as ever, good hygiene is virtue that is its own reward.