Resolve HashTable test case 4
This patch provides important coverage for test case 4,
the basic accessors, by increasing the range of test data
to cover the default data set used in other test cases,
and to test with a broader range of load factors.

Additionally, proper test coverage is provided for all of
the accessors, around half of which were not yet properly
tested.

Finally, as a note for performance, the most extreme case
of a low 'maxLoadFactor' was dropped as it had a catastrophic
impact on test running times.

In a similar vein, support for testing hasher or comparator
functions taking their argument by non-const reference is,
for the immediate future, disabled as the signature of one
of the less frequently used basic accessors is not compatible
with such functions/functors.  Given our lack of interest in
such types other than for standard conformance, and that the
tricky accessor is not directly needed for standard support,
we elect to retain the current interface and drop the test.
Ideally we would like to retain the rest of the test in this
case, but it is still quite difficult to extract the one
function under test from a package when testing something
like basic accessors, where around 10 functions are actively
under test in the one test case.