Final update to HashTable test driver (for now)
This patch must suffice as the end-point for the current work, although the
test driver is still not 'done'.  In this update, we expand coverage of the
equality-comparison operators test, turning up and fixing a test-driver bug
along the way.

The default test data set is expanded to add 'infinity' to the end of the
standard 'maxLoadFactor's to test with.  More test cases are updated to
work with the default (global) tables of test data, and the testing of
extremely low 'maxLoadFactor' values is deferred by default, unless
explicitly requested by defining a macro at compile time.  This is necessary
as it turns out that this one test value was responsible for an excessively
large portion of the test driver running costs.

Attention was paid to try to reduce redundant work, or optimize running of
the most expensive test drivers, but more still needs to be done.

Probably the most important missing work is rounding out the final set of
test cases, to validate as yet untested methods.  This would be a pressing
concern if these methods were not already being indirectly tested by the
four standard unordered containers that depend upon them.

The next pressing task in terms of missing testing is to use the groundwork
laid to test hash and comparator policies that can throw.  This is not as
simple as originally planned though, as we grant only read-only access to
the functors once installed into a HashTable, but need modifiable access
to toggle the throwing behavior.  The regular associative containers
handled this kind of issue by making the toggle into external state,
global to the test driver.  We would prefer to avoid that direction if
possible, but it may be acceptable within the confined of a .t.cpp file
that has its own 'main'.

The final pressing task is to properly document the concerns and test
plans for each test case, and again separately for the template
parameter concerns with a separate test plan for this dimension of
the test driver, and of course to revisit and update the component-
level test plan.  For the parts of the test driver that have now
been implemented, there is some form of this documentation in place,
of varying degrees of rigor, and not always up to date with the
final tested set of concerns.

Given the negative tone of what is missing, it should be mentioned
that the features that are tested are now tested to a high degree
of confidence, which is at least the basic value-semantic operations
of this class.

For future development, there are still a few classes that may
deserve promoting to components in their own right.  Most notably,
the class template 'TestDriver_ForwardTestCasesByConfiguation'
used for dispatching templated test cases for configurate test
case runners will be important to generalizing the idea of testing
templates of more than one template parameter through the 'bsltf'
framework.