Remove Duff's Device from BDE
This is an old deprecated component that has no external
consumers remaining, so eliminate the component before it
accidentally sees reuse.

We have one internal component in BDE making use of this, and
it uses only one method, so move that one peice of code directly
into 'bdeci_hashtable' (itself a deprecated component) as an
implementation detail.

If we were doing more than the barest minimum work to remove
the component, we would update the 'bdeci_hashtable' component
to use 'memset' or 'bslalg::ArrayPrimitives' instead, but that
is left as a future exercise should this component see any
expected further use at all.