Eliminate invented legacy
A small number of bsl components created a 'bde' compatibility
name to reflect their legacy before migrating down into the
'bsl' package, and then created additional 'compatibility'
names when new naming conventions were applied across the whole
of 'bsl'.  That additional layer of new-convention-for-old-names
is confusing an un-necessary.  This patch aims to remove it.

So far, the only two components identified as being affected are
bsls_assert and bslsl_platform.