User defined literals in C++03 'bae' code
This adds whitespace where necessary to fix C++11 parsing issues
when existing C++03 code concatentaes string literals, where one
of the literals is represented by a macro.

Note that this is not a complete patch - we delierately do NOT
address the 'baexml' package, which has similar issues, as it
is too likely we would trample over some ongoing work in this
area.