Remove support for hash<const char *> as string
This patch removes the defintion of the explicit class template specialization
of 'hash' that would treat 'const char *' arguments as string values, rather
than as pointer values.  Attempting to use this template instantiation will
be a compile-time error, prompting users to examine any remaining code using
the old behavior and make appropriate changes.  If, however, they are using
the component correctly and expecting to hash the pointer value, there is no
alternative component to replace their usage with - this is the intended
long term solution.  In that case, the user can define a build macro to
re-enable the dispatch to the general 'TYPE *' partial specialization.