Catch outstanding Microsoft warnings
This patch adds a couple of pragmas to test drivers, after the final
#include, to silence warnings about creating obscure types in the
type system, such as const-qualified function and reference types.

In addition, any of the test drivers using 'printf' for diagnostics
that were not already ported to using 'bsls_bsltestutil' for the
standard test features were duly ported.  No effort was made to
upgrade test drivers still using the standard iostreams library,
which would be more work, but also needs to be done as a matter of
levelization violation.