Correct line numbers used for assertions in test driver
Somehow managed to commit without noticing this test driver fails due to
the recent edits changing line numbers, which break the expected line
numbers in various asserts.

As we have to do further surgery on this test driver, note that is is a
very simple driver, and so swap out the iostreams based reporting for
the expected bsls_bsltestutil version.