Remove bcefX_vfunc6 components
This patch removes the old, long deprecated, vfunc6 components from
'bce' as part of a long-term plan to gradually eliminate these
legacy components.

Note that 'bcefu_vfunc6.h' shows up as an #include in Robo, a
number of times, but there is no sign of any identifiers from
this component itself being used.  Therefore, we retain the
header for the purpose of not breaking backwards compatibility,
and supply the additional #includes for headers that might be
transitively included through it.  However, the header itself
is gutted and devoid of content.  The test driver is a simple
return statement.