Add explicit dependencies to metadata
Adding the previous transitively included headers to 'bcefu_vfunc6'
has introduced an explicit dependency on 'bdes' where previously it
had been a transitive dependency (for the metadata).  This requires
declaring the depency in the package .dep file.

Fixed the presentation of related .dep files while auditing this
issue, which is alphasorting the dependant packages, and one
package per line.