Remove use of mallocfree allocator from header
At a very high level, we have Bloomberg code that defines macros named
'malloc' and 'free' in order to substitute different allocators using
shared memory.  This breaks headers that rely on expecting to use
'malloc' and 'free' from the C library.

HashTable uses the malloc-free allocator only to supply untracked
memory to a data structure validation routine in SAFE_2 builds.  It
is important that this is neither the global, default nor object
allocator as that would change the expected results for memory
consumption in the test driver.

The simple patch employed here is to move the call to the
malloc-free allocator into the component .cpp file, and provide
a static member function in one of the utility classes to
supply it where needed inside our template definition, and so
allowing us to move the problematic header out of our own
header, so avoiding issues with other builds.