Clean-up bcema_SharedPtr test driver
This patch offers two key clean-ups.  First, it eliminates
the complete copy of the original sharedPtr implementation,
which was hanging around only to support a single test case
with a negative test number, so it was not even being run
regularly.

Secondly, ran the bdespell and bdeflag tools, and cleaned
up the majority of the issues reported.

There should be no functional change of behavior as a
consequence of this patch, although the test driver
compiles a few seconds faster.