shared pointer should support construction with a deleter and no allocator
This patch applies the basic resolution for DRQS 3715472, but
does not add any new test cases (yet).  The code is copied over
from a branch that jointly resolved both this issue, and the
issue of constructing a shared pointer with a native pointer
and no allocator, where we must change our semantics to use
a deleter that simply call delete, rather than use the
default allocator to destroy the shared object.

By removing the work for the change of deleter, the test
driver will need a different set of patches to validate
the new behavior - currently we are testing just that we
did not break any existing code.  Also note that the test
driver for shared pointer is little more than an extended
breathing test, so it is not clear what level of testing
is appropriate for this specific patch - we will probably
try to scrape by with a similar level of breathing test
for the new range of constructor overloads enabled by
this patch.