Simplify SharedPtr to support const void (DRQS 33549823)
There is an outstanding request to extend the dummy reference type
used to transparently support compiling 'operator*' for void types
to also support 'const void' types.

In practice, we already have a generic version of this reference
aliasiing component in the type traits facility, and it already
has the desired additional support for cv-qualified void types.

Hence, the simple fix is to eliminate the component-private
workaround type and simply use the (more recent) type-trait.
Majority of the work is extending the test driver to validate
this case.

Note that we still do not support volatile-qualified 'void'
types, as we do not support volatile qualified pointers in
general.  The test case exposing this is retained but
disabled, in case we want to extend such support in the future
as part of our C++ Standard compliance effort.