Fix sharedptr to not use deprecated managedptr functionality
This issue reflects a patch to managed pointer a few BDE releases ago
to avoid accidental reliance on undefined behavior that happens to do
just the right thing on all the platforms that we test on.  This is a
picky patch, but technically an improvement, and the kind of issue
that might get picked up and commented on once we open source this
component.