Remove dangerous SharedPtr constructor for test build
This patch experiments with removing the dangerous constructor for
'bcema_SharedPtr' taking a single 'bcema_SharedPtrRep *' pointer.
The danger is that if the rep is from an aliased shared pointer,
the user's code will have undefined behavior that is not clearly
spelled out in the contract, and is hard to test and avoid.  The
simple solution is to always call the constructor that also takes
a pointer to the object being aliased, even when no aliasing occurs.

It is believed that this safe usage pattern is the only one employed
in BDE code, so this patch simply makes the dubious constructor
private, removes the definition of this ctor with an #if check, and
fixes up the test driver for this case to simply use the preferred
constructor.

After manually checking that 'bde-core' builds, this change-set will
be used to validate that the whole of bde, including bde-bb, builds
correctly, in which case a better documented patch will be submitted
for the next BDE release.