Remove dangerous ctor from bcema_SharedPtr
The explicit single-argument constructor taking a pointer to
a 'bcema_SharedPtrRep' is difficult to use correctly, as the
original shared pointer may well be aliasing a different
target than that reported by the rep itself.  This constructor
offers no functionality that is not available through the
similar constructor that also takes a pointer to the target
object, its only purpose is a dangerous ease of use.

This patch removes that constructor, initially making it
private, but not defined, in order to catch any lurking
users of this constructor that we are not aware of.
Initial testing suggests that there are no such users.

The private constructor is documented with a brief
description of the problem, and guidance to use the
preferred constructor instead.  This constructor
may be removed in a future BDE release, for example
when we open source this component by moving it
down to 'bsl'.