Windows workaround for test drivers of unix headers
We have a couple of components that wrap standard POSIX
headeers that do not have a Windows equivalent.  It is
preferred for the component header to fail explicitly
on Windows when it tries to include the platform
header, rather than silently succeeding, only to fail
more subtly later when the user tries to make use of
the features declared in that header.  Hence, this
patch ammends the test drivers to perform a quick
platform detection before trying to include the
headers under test, avoiding the problem on Windows.