Support function pointer deleters for SharedPtr on Windows
The nightly test results revealed a nasty compiler bug on Windows.
When we pass the address of a function directly to the constructor
of 'bcema_SharedPtr', MSVC treats the results pointer value as if
it were a function reference, and so we fall into the as-yet
uncoded logic to support deducing the DELETER as a simple function
type, rather than dispatching as a pointer.

In order to get proper support for function poninter deleter on
MSVC we therefore go the rest of the way, and complete support
for function types and function references as deleters too.

Most of this work is actually in the shared pointer out-of-
place representation class, which gets a drive-by spring
clean on some style guide issues.