Disable a couple of tests on AIX
The IBM compiler cannot handle the type deduction when a
function is passed directly as a deleter to a 'SharedPtr'
constructor, and insists that we explicitly take the
address.  We simply disable this test case on AIX as
it is a corner of the library that we do not care to
use in practice, but support only for standard
conformance.