Catch exception escaping bsls_assert test in SAFE_2
When testing for SAFE_2 builds, the usage example features throwing
a 'std::exception' object, that is not caught in the test case.
This is a simple patch to explicitly insert a try/catch block unless
building in a mode where exceptions are disabled.  We explicitly
check the build mode flag and use try/catch directly, rather than
using the 'bsls_exceptionutil' wrapper, as that component directly
depends on 'bsls_assert' already, and we must not introduce a cycle.