shared pointer support for writing to arbitrary streams
This is a requirement for one day migrating 'bcema_SharedPtr' down to
the 'bsl' namespace as our drop-in implementation of the C++ standard
shared pointer type.

The C++ standard shared pointer supports writing to an arbitrary
instantiation of the 'basic_ostream' template, so we must generalize
our support from streaming only through the 'ostream' typedef.