Support all logical comparisons for bcema_SharedPtr
In order to support a future migration to 'bsl', 'bcema_SharedPtr' should
support the full standard shared_ptr interface, and that includes the full
set of comparison operators, both with other shared pointers, and with
null pointer literaels.

This patch has the initial implementation, but no test driver, for the new functions.