Cleanup of various small issues
This patch provides a spring clean to 'bcema_sharedptr' in
anticipation of open sourcing.  The component-private
implementation of the unspecified-boolean type idiom is
replaced with the general facility from 'bsls', saving on
the testing concerns.

Redundant entries are removed from the component-level
checklist in the test driver.

Many small methods are explicitly marked as 'inline'.

Any methods defined inline in the class are moved
out-of-line, barring a strong reason (compile-fail)
to return them.

Un-needed headers are moved into a deprecated section
guarded by a test for the BDE_DONT_ALLOW_TRANSITIVE_INCLUDES
macro.

Minor style-guide issues on whitespace are also
addressed.