Restore private bcema_SharedPtr constructor for one release
To allow time for the existing consumers of the dangerous
constructor to update their code, revert the changes for
DRQS 35875366 for one release.

This patch restores the public constructor (without
explicit test cases, the old tested code is copied
into place) and documents it as deprecated with the
recommended replacement.