Rename SharedPtr template parameter to ELEMENT_TYPE
Unfortunately there are a number of places in the full Bloomberg
code base that define a macro named 'TYPE', and so break when
included before the 'bcema_sharedptr' header.

This patch avoids the issue by renaming 'TYPE' to 'ELEMENT_TYPE'
and updating all documentation accordingly.  A small number of
additional doc corrections are made in passing, relating to
other recently merged branches invalidating a few notes.