Replace use of 'OTHER_TYPE' as identifier
One of the IBM AIX platform headers defines a macro named
'OTHER_TYPE' so it is not safe to use this as a template
parameter name.  This patch replaces 'OTHER_TYPE' with a
more specific name, either 'ANY_TYPE' or 'COMPATIBLE_TYPE'
depending on the context of the function it is parameterizing.