Initial implementation of 'throw' utility
There is a need for a component to wrap the throwing of various
native exception classes specified by the language, and accessed
below the 'bslstl' and 'bsl+bslhdrs' levels - for example,
allocators will want to indicate failure by throwing a standard
bad_alloc exception.  It is important that these classes truly
are the native platform exception classes, and not a BDE wrapper
of said class, to ensure link-time compatibility with other code
catching the standard exceptions.  Otherwise, we end up duplicating
the standard exception hierarchy not only in BDE, but also in
every exception handler trying to catch a standard exception.