Initial migration of ManagedPtr to bslma
This patch migrates all the bdema_manageptr components into
bslma, and ensures that the components, but not the test
drivers, build in this new location.  All names and doc
should have been updated, but there probably remains a fair
ammount of bookkeeping and tidying to catch up on.

There are also a small number of levelization violations that
have not yet been addressed.  Notably, at least one component
supports streaming with the standard iostreams, and morea
awkwardly, the 'release' member functon of 'ManagedPtr'
returns a 'bsl::pair'.  An alternate version of this API has
been implemented using an out-parameter to avoid violating
correct levelization, but the problem of compatibility with
existing code remains.