Add weak_ptr constructor to shared_ptr
This additional constructor is mandated by the C++11 standard.  Note that
it is also required to fail by throwing an exception if the weak_ptr has
expired, a tweak that we have not yet implemented.

In addition, flag the redundat BDE-named functions as deprecated/internal
so that they do not escape into the bsl-oss repository.  There is no
value in creating a naming confusion for standard methods.