Updated behavior for constructing owned pointer without an allocator
This set of changes is a pretty big deal, so guarded with a few
feature-macro checks so that we can validate the extent of
compatiblity issue before removing those feature macros for the
production release.

The essence of the change is that, in order to meet the contract
of 'shared_ptr' in the C++11 standard, passing a pointer to an
object to be owned without also passing an allocator will cause
that object to be destroyed, at the end of its managed lifetime,
by a call of 'delete ptr;'.  This change should not impact the
large majority of users who never change default allocator, and
do not overload the class-specific 'operator delete'.  The latter
group will now get a feature they have been requesting for some
time, but may be actively working around in their code.

This change has no effect on any construction of a smart pointer
where either a deleter or allocator is explicitly supplied, even
in the case of passing '0' to request the default allocator.

There is a matching change to the 'shared_ptr::load' function,
where we provide an additional overload rather than defaulting
the allocator argument.