Implement 'get_deleter' for shared_ptr
This implementation adds a virtual function to the (non-template)
base class 'getDeleter', whihch is *non* const, and takes a
'std::type_info' object by const reference, that the derived
implementations can use to check against if they support a
deleter.  This method is deliberately not const-qualified in
order to all in-place representations returning data members
without worrying about const-casts and undefined behavior.

The inplace representation does not support deleters, so must
always return a null-pointer.  For the out-of-place rep, we
are dealing with a template that knows the type of its
deleter, so we defer the typeid call until someone actually
tries to call 'getDeleter'.  That makes this addition
essentially no more expensive for a client that never uses
the new facility (most of them) at the cost of performing
two 'typeid' calls to retrieve the type info when evaluating
the call.  It is not clear that storing the type info in a
data member would be more efficient anyway, so this seems
clearly the preferred approach.

We also rejected the idea of adding a method to return the
type info for the deleter and asking the standard function
'get_deleter' to do the comparison.  In that case, we
would have had to determine a deleter type to return for
types that do not support deleters, such as the in-place
representation.  That did not seem practical, and would
add two virtual functions to the vtable rather than one,
to save the cost of passing a type info object by
reference.  Again, a cler win (we think) for the chosen
implementation.