Complete patch for gcc 4.8.1 warnings
This patch addresses two issues with warnings.  The
first is that gcc 4.8.1 introduces a new warning for
unused 'local' typedefs, that are typically stale code
that should be cleaned up.  However, that is not true
in all cases that the warnings flags for, and in fact
we must produce a new implementation for 'BSLMF_ASSERT'
using function declarations rather than typedefs to
achieve the same effect.  Note that testing the error
message reported shows the exact same message with
both implementations, so there should be no new QoI
issues raised with this implementation.

A second issue noted and fixed is a large number of
warnings from 'BSLMF_ASSERT' expressions in 64-bit
builds, as the expression type of the static assert
construct is 'int' rather than 'bool', and the gcc
compiler likes to warn about conversions potentially
changing values with the '!!' idiom potentailly
producing a large value to truncate.  This issue is
most easily solved by using 'bool' as a more natural
match for the type of a boolean expression, which
happens to avoid the narrowing concerns highlighted
by the compiler.