fix for asserttest component and no exceptions
The bulk of this patch is a mechanical edit of much duplicated code
that will stop testing 'ASSERTTEST_ASSERT_PASS' type expressions
with values that expect them to fail, unless exceptions are
available.

There is no issue with the 'fail' macros, as they must always
expand as empty unless exceptions are available - so we do NOT
want to disrupt successful testing of that feature.