Initial patch to eliminate old-style traits
This patch addresses the issue of components that were added
contemperaneously with the switch to new-style type traits.
These components are the only remaining use of the 'bslalg'
facility in 'bsl' and should be eliminated so that we can
fully deprecate the facility, ideally migrating it to
'bsl-internal'.  The latter requires the whole of BDE to
be legacy-traits free as well though.