Working implementation of managed pointer
This patch layers 'bdema_ManagedPtr' on top of 'bslma::ManagedPtr'.
The main issues to be aware of are that private methods designed
to interfere in overload sets to not play well with 'using'
declarations, so we end up re-implementing those methods to
directly forward the calls to base.

For the test driver, many 'ASSERT_SAFE_FAIL' need to become
'ASSERT_SAFE_FAIL_RAW', as the assertion is now triggered in
a base class method, due to the aforemention using declarations.
The obvious exceptions here are the 'load' functions.

Finally, noticed that one of the constructors was asserting
unexpectedly, and this turns out to be because the base class
performs the transfer of a passed managed pointer, the assertion
that *that* pointer object not be null would no longer be true.
We must instead rely on the equivalent assertion in the base
class constructor.