Fixes for checkin
This change-set has numerous test driver fixes, catching the
differnce between C++11 semantics for some functions, and the
legacy BDE semantics.

It also applies the patch to warn if calling a constructor
whose semantics will be changing in the following release.

At this point, there is still a fair amount of arithmetic
to patch in the shared pointer test driver, but otherwise
this should now be ready to merge to master - hence the
hasty check-in tonight in order to get some testing with
the nightly regression tests on the rest of the code.