Fix MSVC specific error
Visual C++ accepts partial specializations on cv-qualified
pointer types (where the pointer, and not the pointed-to
type, is cv-qualified) but it cannot match them when
given such a type.  The workaround is to match any
cv-qualified tyoe, which is a broader match than necessary
and may cause an extra few template instantiations in
rare cases, but probably not so that we will notice.