Fix assertion in basic_string destructor
We have the following invariant check in the 'basic_string' destructor:
       BSLS_ASSERT((*this)[this->d_length] == CHAR_TYPE());
This invariant, or something similar to this, is guaranteed by C++11.  The
problem is that 'CHAR_TYPE' can be any POD type, and the definition of POD does
not mandate an operator== for arbitrary PODs.  This is the cause of the Apache
failure.
Instead, we should use the char traits template parameter to confirm that the
null terminator compares equal to the value-initialized POD.  This is confirming
 a weaker property than the standard explicitly guaranteeing that the value *is*
 CHAR_TYPE(), but has the benefit of compiling for arbitrary types.  If we were
super-keen, we might write some crazy meta-programming to use memcmp if the type
 is bitwise-equality-comparable, and delegate to the char traits otherwise, but
we would pay for that in every compilation, as this is the basic_string
destructor and not a sanity check in the test driver.