std/bslhdrs - declare only relevant names
Components in 'bsl+bslhdrs' and 'bsl+stdhdrs' have bent to the will of
backwards compatibility and declared names that do not exist in the underlying
standard C++ headers.  This causes issues in strictly standard-compliant
headers since name declarations may clash or may not be found.

This commit only declares the appropriate names, preserving the old behavior if
the 'BDE_OMIT_INTERNAL_DEPRECATED' define is present.