Better support for C++11 allocators in shared pointers
This patch more thoroughly addresses and tests the concerns of the standard
'allocate_shared' API.  One unexpected consequence is that allocator_traits
'construct' method must now support overloads up to at least 14 arguments,
while the current tested version supports only 5.  The script used to
generate the overloads from the basic variadic code was re-run, so although
the test driver has not been extended (yet) we have confidence from previous
tests and the specific testing of 'allocate_shared' functionality that this
works correctly.

gcc has a compiler bug with the variadic template implementation that strikes
when supporting 'allocate_shared' overloads for BDE-style allocators, which
is a regression from gcc 4.4, and remains present up to the latest gcc
available on our build machines, which is v4.8.1.  Luckily there is a
reasonably simple workaround using 'enable_if', although it would be much
nicer to not require it.

An additional clean-up pass is still required, now that the basic
functionality has finally landed, along with better testing that the
supplied allocator is passed to the members of the shared object,
when appropriate (and not otherwise).