Clean up 64-bit warnings in hash table
This patch clears up just under 1300 warnings when building the hash
table test driver on a 64-bit platform.  Most of the warnings arise
from using a test allocator which deliberately supports only a 32-bit
address space, i.e., has a narrower 'size_type' than 'std::size_t'.

We do not yet have a test allocator to verify behavior when 'size_type'
is *wider* than 'std::size_t', but that appears less useful for this
specific data structure, as the standard hash-function requirements
include that the function returns a 'std::size_t', so even if we can
allocate more memory, we literally cannot address more buckets due to
the granularity of the hash function.