Do not test deprecated features unless supported
When the macro BDE_OMIT_INTERNAL_DEPRECATED is defined, deprecated
functions are removed from legacy interfaces.  This patch ensures
that the test drivers continue to build and pass when this macro
is defined.  That corresponds to the state of the code that will
be pushed to the open source branch.