Address more bde_verify warnings
In some cases we are using #pragma bde_verify to turn off warnings.
Where such use is wrapping a usage example, it is probably intentional
and will remain to final release.  In other cases, it is mostly a way
to reduce clutter to help focus on fixing a single issue at a time,
and such pragmas should be removed before the branch is released.