Restore earlier usage example edits for bslmf_enableif
It appears an eariler round of doc fixes updated the usage examples for 'bslmf::enable_if',
but did not make corresponding changes to the test driver.  The last round of edits fixed
up issues in the test driver, and then overwrote the whole usage example in the header,
losing the previous fixes that had no propagated back to the test.  This showed up in the
phabricator review.

This change set tries to merge the best of both worlds from both sets of edits, and to
bring the .t.cpp and the .h usage examples back into synch.