First pass at cleaning up shadowing warnings
This patch takes a first stab at the shadowing warnings,
specifically transforming 'allocator' to the preferred
'basicAllocator' for all of the standard containers.
Additional passes are still required for further type
shadowing issues, related to 'deque', 'hash' and
'string', plus on more pass at 'allocator_traits' where
a new synonym for 'allocator' may be desired for the
'allocate' and 'deallocate' functions.