Start refactoring shared_ptr test driver
This refactoring starts to reorder the test cases, based on testing as
an in-core value semantic type, rather than as a more general mechanism.
This brings a bit more order to the testing, and exposes many of the
gaps.

As part of updating the tests, add missing asserts for narrow contracts
in the component header, and update contracts relating to BDE factories
that may also be allocators.  Testing revealed a more complex behavior
than documented in the contract, and it is believed that the current
behavior is intended.