Change default build type to 64-bits on 64-bit systems.
This patch changes the mechanism used to determine the system's processor type
from various mechanisms to Python's "platform.machine()" library function. It
uses this information to change the default for the --abi-bits parameter to
the 'configure' command.

In addition, it separates the code that determines this information from the code
that determines the compiler type and version, as those are unrelated operations.