Second pass at cleaning up shadowing warnings
This patch deals with the remaining name-shadowing warnings
reported when building the BDE stack in VC++2013, but not
building any of the test drivers.

We specifically opt to make no changes to hash table or
shared_ptr due to other branches merging in the near future making extensive changes to these components, so
any warnings remaining in this area will be much simpler
to clean up after those branches land.

There are additional warnings still reported when building
with VC2008, which also builds all of our test drivers,
but those will be handled in a separate pass.