Fix errors renaming parmameters in string functions
This patch fixes the errors introduced renaming several parameters in
'basic_string' functions to avoid shadowing type-names, which raises
annoying warnings on Windows /Za builds.  Unfortunately the original
fix did not refactor properly, and the errors were not noticed by
the MSVC compiler when building the .cpp files only, and not any of
the .t.cpp (which is an odd thing to do, but makes sense in my
build environment).