Final fixup of formatting issues
This is a final pass at patching up the simplest of formatting issues
highlighted by 'bde_verify', not all of which were introduced as part
of the current refactoring, but are best cleaned up at the same time
that these files are already being editted.

This ignores the majority of warnings issues by 'bde_verify' which
would involve a more substantial reworking of many of these files,
such as correctly alphasorting declarations and definitions.