Add missing headers
The latest BDE can be much stricter about exporting
names from standard headers, and when built in such
a mode, transitive dependencies in the decimal code
are revealed.

As a drive-by fix, alphabeticize the headers in any
file that is touched, in order to insert the missed
headers at the right place.