Correct use of 'specified'
This change set adds (most) missing uses of 'specified' in function
contracts, and removes excessive usage, as highlighted by 'bde_verify'.
A small number of issues remain, where the fix would involve
substantial rewording of the function contract, which is more work
than I want to spend on this right now.  Similarly, there is no
attempt to divine meaning and supply documentation for the many
function parameters that are flagged as undocumented.  That seems to
be an important issue that should be picked up, but will involve
spending a little more time to actively research and document each
parameter - this pass is mostly mechanical.