Run script to construct with 14 arguments
Shared pointer factory functions need to call 'allocator_traits'
'construct' method with up to 14 arguments, for compatibility
with existing code prior to availability of variadic templates.

Pulling out this extension into a separate change-set as it has
frequently been the soure of merge conflicts during the long
lifetime of the shared pointer branch.