Fix casts for IBM
The current production IBM compiler cannot handle 'remove_pointer'
for function pointers, which in turn leads to casts that cannot
be easily broken down into a sequence of 'const_cast' and
'reinterpret_cast' in our generic code.  Therefore, we fall back
on the pre-existing solution of a hard C-style cast, which was
essentially the purpose of this patch (satisfying bde_verify)
for all other platforms.