Add a test case for vector of function pointers
This patch restores the test supplied by Andrew Lin that demonstrates
that vector of function pointer fails to compile.  This is applied to
a clean branch updated from master, as there has been substantial
churn in the 'bslalg_arrayprimitives' component that is trickier to
resolve at the merge level than to recreate the final edits on this
new branch.  Applying edits in stages, and the 'vector' test is a
clean merge.